+++ src/componements/WordleLogo.js
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export default function WordleLogo() {
+  return (
+    <div className='flex flex-col gap-2'>
+          <div><p className='text-4xl text-center	tracking-[0.5em]'>WORDLE</p></div>
+          <div><p className='text-4xl text-center	tracking-[0.5em]'>GAME</p></div>
+        </div>
+  )
+}
